+++ Public/Remove-PhoenixPackage.ps1
@@ -34,7 +34,9 @@ function Remove-PhoenixPackage {
         )]
         [ValidateSet(
             'WinGet',
-            'Chocolatey'
+            'Chocolatey',
+            'Scoop',
+            'MSI'
         )]
         [string]$Provider
     )